tests/wpt/tests/css/css-flexbox/align-items-baseline-column-vert-rl-table-item.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#baseline-rules">
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#cross-alignment">
<link rel="author" href="mailto:sammy.gill@apple.com" title="Sammy Gill">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
#target {
  display: flex;
  flex-direction: column;
  width: 200px;
  align-items: baseline;
  border: 3px solid black;
  font-family: monospace;
  font-size: 10px;
  line-height: 10px;
}
#target > :nth-child(1) {
  background: hotpink;
  writing-mode: vertical-rl;
  padding-left: 30px;
  margin-left: 10px;
}
#target > :nth-child(2) {
  background: papayawhip;
  writing-mode: vertical-rl;
}
.inner {
    border: 5px solid black;
    padding: 5px;
}
</style>
<body onload="checkLayout('#target > *')">
<div id="target">
  <table class="inner" data-offset-x="140">
    <tr>
      <td style="vertical-align: baseline;">
        <div>two<br>lines</div>
      </td>
    </tr>
  </table>
  <div data-offset-x="188">hello</div>
</div>
</body>
