LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-items-baseline-column-vert-rl-table-item-expected.txt

two
lines
hello

FAIL #target > * 1 assert_equals:
<table class="inner" data-offset-x="140">
    <tbody><tr>
      <td style="vertical-align: baseline;">
        <div>two<br>lines</div>
      </td>
    </tr>
  </tbody></table>
offsetLeft expected 140 but got 139
PASS #target > * 2
